UI improvements and refactoring
