+++ templates/vue/src/components/modals/ExportDuplicateModal/DuplicateTapestryButton.vue
@@ -1,21 +1,25 @@
 <template>
   <div>
-    <b-button
-      block
-      data-qa="duplicate-tapestry-button"
-      variant="primary"
-      @click="duplicateTapestry"
-    >
-      <b-spinner v-if="loading" small data-qa="spinner"></b-spinner>
-      <div v-else>Duplicate on this site</div>
-    </b-button>
-    <b-alert :show="showConfirmation" variant="success" style="margin-top: 1em;">
+    <b-form-text class="mt-2">
+      Create a duplicate copy of this tapestry on this site.
+    </b-form-text>
+    <b-alert :show="showConfirmation" variant="success">
       Your new Tapestry is ready! Click on the link below to view it.
       <br />
       <a data-qa="duplicate-tapestry-link" :href="link" target="_blank">
         {{ link }}
       </a>
     </b-alert>
+    <b-button
+      block
+      data-qa="duplicate-tapestry-button"
+      variant="primary"
+      class="mt-3"
+      @click="duplicateTapestry"
+    >
+      <b-spinner v-if="loading" small data-qa="spinner"></b-spinner>
+      <span :style="loading ? 'opacity: 50%;' : ''">Duplicate on this site</span>
+    </b-button>
   </div>
 </template>
